x1plusd: Dynamic loading of i2c drivers
